+++ core/src/main/scala/hammerlab/iterator/package.scala
@@ -5,6 +5,7 @@ package object iterator
      with end
      with iterator.either
      with group
+     with level
      with range
      with sample
      with scan
